% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateLorenz.R
\name{integrateLorenz}
\alias{integrateLorenz}
\title{Integrate the Lorenz Equations}
\usage{
integrateLorenz(x = 0.01, y = 0.01, z = 0.01, sigma = 10, rho = 28,
  beta = 2.6666666666667, stepSize = 0.01, stepsPerSample = 100,
  numSamples = 1000, numSamplesBetweenInteruptChecks = 100)
}
\arguments{
\item{x}{initial value of x}

\item{y}{initial value of y}

\item{z}{initial value of z}

\item{sigma}{value of sigma (10 default)}

\item{rho}{value of rho (28 default)}

\item{beta}{value of beta (8/3 default)}

\item{stepSize}{size of an individual intergration step (0.1 default)}

\item{stepsPerSample}{number of integration steps per sample (100 default)}

\item{numSamples}{number of samples to collect (1000 default)}

\item{numSamplesBetweenInteruptChecks}{the number of samples collected
between each check for user interupts. A small number will slow down the 
integration, while a large number will inhibit the user's ability to stop a
long running integration. The default is 100 samples.}
}
\value{
results of the integration as a matrix of numSamples rows and x, y, z
  columns
}
\description{
Taken from wikipedia: [Lorenz
System](https://en.wikipedia.org/wiki/Lorenz_system)
}

